Assorted compile fixes from Jason Sundram, thanks!
git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@3225 0e98fd9d-079e-457f-9c5c-d482923f975d

+++ src/apps/sfplugin/sfplugin.cpp
@@ -176,10 +176,10 @@ loadOptions()
 {
 	helpopt = cmd_options.getBoolOption("help");
 	usageopt = cmd_options.getBoolOption("usage");
-	start = cmd_options.getRealOption("start");
-	length = cmd_options.getRealOption("length");
-	repetitions = cmd_options.getRealOption("repetitions");
-	gain = cmd_options.getRealOption("gain");
+	start = (float)cmd_options.getRealOption("start");
+	length = (float)cmd_options.getRealOption("length");
+	repetitions = (float)cmd_options.getRealOption("repetitions");
+	gain = (float)cmd_options.getRealOption("gain");
 	pluginName = cmd_options.getStringOption("plugin");
 	fileName   = cmd_options.getStringOption("filename");
 	loop = cmd_options.getBoolOption("loop");
